# docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.TokenPair:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
    type: object
  errors.Code:
    enum:
    - 0
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    - 1006
    - 1007
    - 1008
    - 2001
    - 2002
    - 2003
    - 2004
    - 3001
    - 3002
    - 3003
    - 3004
    - 4001
    - 4002
    - 4003
    - 4004
    type: integer
    x-enum-varnames:
    - Success
    - ErrBadRequest
    - ErrUnauthorized
    - ErrForbidden
    - ErrNotFound
    - ErrMethodNotAllowed
    - ErrConflict
    - ErrTooManyRequests
    - ErrInternalServer
    - ErrInvalidParams
    - ErrBindJSON
    - ErrBindQuery
    - ErrBindForm
    - ErrDatabase
    - ErrRecordNotFound
    - ErrRecordExists
    - ErrRecordInvalidate
    - ErrTokenInvalid
    - ErrTokenExpired
    - ErrTokenMissing
    - ErrPermissionDenied
  handler.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  handler.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  handler.RegisterRequest:
    properties:
      email:
        type: string
      nickname:
        maxLength: 50
        type: string
      password:
        minLength: 6
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  response.Response:
    properties:
      code:
        $ref: '#/definitions/errors.Code'
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/cccvno1/nova
  description: |-
    Nova 是一个基于 Go + Echo 的现代化 Web 应用框架，提供用户认证、RBAC 权限管理、文件上传、异步任务队列、审计日志等功能。

    主要特性：
    - 用户认证与授权（JWT）
    - RBAC 权限管理（基于 Casbin）
    - 文件上传与管理（支持本地/OSS/S3）
    - 异步任务队列（基于 Redis）
    - 定时任务调度（基于 Cron）
    - 审计日志（自动记录所有操作）
    - 限流保护（IP/用户维度）
    - 健康检查与监控
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Nova API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: 使用用户名和密码登录，返回访问令牌和刷新令牌
      parameters:
      - description: 登录信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功，返回访问令牌和刷新令牌
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth.TokenPair'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 用户名或密码错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户登录
      tags:
      - 认证
  /auth/logout:
    post:
      consumes:
      - application/json
      description: 登出当前用户，将访问令牌加入黑名单
      produces:
      - application/json
      responses:
        "200":
          description: 登出成功
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权或令牌无效
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 用户登出
      tags:
      - 认证
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: 使用刷新令牌获取新的访问令牌
      parameters:
      - description: 刷新令牌
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 刷新成功，返回新的访问令牌
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 刷新令牌无效或已过期
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 刷新访问令牌
      tags:
      - 认证
  /auth/register:
    post:
      consumes:
      - application/json
      description: 创建新用户账号
      parameters:
      - description: 注册信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 注册成功，返回访问令牌和刷新令牌
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth.TokenPair'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: 用户名或邮箱已存在
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户注册
      tags:
      - 认证
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"